fix: secure YouTube hostname validation (#91) (thanks @RinZ27)
