+++ apps/chrome-extension/src/entrypoints/background.ts
@@ -20,6 +20,7 @@ import { createDaemonRecovery, isDaemonUnreachableError } from "../lib/daemon-re
 import { logExtensionEvent } from "../lib/extension-logs";
 import { loadSettings, patchSettings } from "../lib/settings";
 import { parseSseStream } from "../lib/sse";
+import { isYouTubeWatchUrl } from "../lib/youtube-url";
 
 type PanelToBg =
   | { type: "panel:ready" }
@@ -389,26 +390,6 @@ function urlsMatch(a: string, b: string) {
     return next === "/" || next === "?" || next === "&";
   };
   return boundaryMatch(left, right) || boundaryMatch(right, left);
-}
-
-function isYouTubeWatchUrl(value: string | null | undefined): boolean {
-  if (!value) return false;
-  try {
-    const url = new URL(value);
-    const host = url.hostname.toLowerCase();
-    if (host === "youtu.be") {
-      const id = url.pathname.replace(/^\/+/, "").trim();
-      return Boolean(id);
-    }
-    if (host !== "youtube.com" && !host.endsWith(".youtube.com")) return false;
-    const path = url.pathname.toLowerCase();
-    if (path === "/watch") return Boolean(url.searchParams.get("v")?.trim());
-    if (path.startsWith("/shorts/")) return true;
-    if (path.startsWith("/live/")) return true;
-    return false;
-  } catch {
-    return false;
-  }
 }
 
 async function extractFromTab(
